Solving all the problems
Setting up jobDisplayNew to send a mail out with an application.
Handled error handling on login.

Set up delete button on Job; works properly now that the button is set
to type=“button” instead of the default which is submit.

Changed the way the model reading for the REST API works. Now it pulls
from an object. Just add it to the object as a property named the name
of the model and put that same name in the array in myApp.js and
everything will work fine. No need to add an if statement!

Made some minor changes hoping to fix regular links, but no luck. We
should just switch to a one page app and be done with it; it’s
shouldn’t be too hard IIRC. and we’ll never have to deal with
dependency or link issues ever again.